[tools] Arrange tflitefile_tool
Let's arrange tflitefile_tool before revision.
