+++ tools/tflitefile_tool/saver/config_saver.py
@@ -14,9 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from operator_wrapping import Operator
-from tensor_printer import TensorPrinter
-from option_printer import OptionPrinter
+from ir.operator_wrapping import Operator
+from printer.tensor_printer import TensorPrinter
+from printer.option_printer import OptionPrinter
 
 
 class ConfigSaver(object):
